Video 4: Lazy Load, Suspense y paginación con React; Crear hook useNearScreen desde cero.
